Fixes long path handling by GitUnite
Previously, long file paths were not handled and ended up with the (cryptic) error that the specified path was too long (even though the original path is not too long).
This is particularly useful with repositories containes node_modules for instance.
This fix is using the "hidden" property `FullPath` which can handle longer file paths.
This is what you find the most used workaround on the internet. One such example : http://stackoverflow.com/a/9589335